pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paimon"
version = "0.1.0"
description = "个人 AI 助手系统 — 神圣规划 (AIMON)"
readme = "docs/aimon.md"
requires-python = ">=3.12"
dependencies = [
    # LLM Providers
    "openai>=2.30.0",
    "anthropic>=0.87.0",

    # Config & Settings
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",

    # Channels
    "aiogram>=3.25.0",
    "qq-botpy>=1.2.1",

    # WebUI
    "aiohttp>=3.9.0",

    # Utilities
    "loguru>=0.7.0",
    "telegramify-markdown>=1.1.1",
]

[project.scripts]
paimon = "main:entry"

[tool.setuptools.packages.find]
where = ["."]
include = ["paimon*"]
